requirements.txt
fastapi
uvicorn
python-dotenv
PyPDF2
pytesseract
pdf2image
Pillow
google-generativeai
